refactor(agentic-sorting): extract useCanvas2D hook (dedupe ResizeObservers)
Legibility pass (4/4, code tidy): collapse the two near-identical DPR-aware
ResizeObserver effects (arena + trajectories canvases) into a single
useCanvas2D(ref, size, onResize, deps) hook. Behavior-preserving.

Stopped short of extracting <ObjectiveFrozen>/<PopulationMix>: the two
objective/frozen call sites aren't identical (sandbox carries the Levin-faithful
hint; the Lab copy is compact and embedded among condition sliders), and
agentsNode is already a single reused node — extraction would add conditionals
without clearly simplifying.
